logos-I/pyproject.toml
[project]
name = "logos-i"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.2",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anthropic==0.49.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "cachetools==5.5.2",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "cryptography==44.0.2",
    "deprecation==2.1.0",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "ecdsa==0.19.1",
    "email-validator==2.2.0",
    "exa-py==1.12.1",
    "fastapi-cli==0.0.7",
    "fastapi[standard]==0.115.12",
    "frozenlist==1.6.0",
    "google-api-core==2.24.2",
    "google-api-python-client==2.167.0",
    "google-auth==2.39.0",
    "google-auth-httplib2==0.2.0",
    "googleapis-common-protos==1.70.0",
    "gotrue==2.12.0",
    "h11==0.14.0",
    "h2==4.2.0",
    "hpack==4.1.0",
    "httpcore==1.0.8",
    "httplib2==0.22.0",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "hyperframe==6.1.0",
    "idna==3.10",
    "inflection==0.5.1",
    "iniconfig==2.1.0",
    "jinja2==3.1.6",
    "jiter==0.9.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.23",
    "langchain-core==0.3.54",
    "langchain-openai==0.3.14",
    "langchain-text-splitters==0.3.8",
    "langgraph==0.3.31",
    "langgraph-checkpoint==2.0.24",
    "langgraph-prebuilt==0.1.8",
    "langgraph-sdk==0.1.61",
    "langsmith==0.3.32",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mcp[cli]==1.6.0",
    "mdurl==0.1.2",
    "motor==3.7.0",
    "multidict==6.4.4",
    "openai==1.75.0",
    "orjson==3.10.16",
    "ormsgpack==1.9.1",
    "packaging==24.2",
    "pluggy==1.5.0",
    "postgrest==0.19.3",
    "propcache==0.3.1",
    "proto-plus==1.26.1",
    "protobuf==6.30.2",
    "pyairtable==3.1.1",
    "pyasn==1.6.2",
    "pyasn1==0.4.8",
    "pyasn1-modules==0.4.1",
    "pycparser==2.22",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "pydantic-settings==2.8.1",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "pymongo==4.12.0",
    "pyparsing==3.2.3",
    "pytest==8.3.5",
    "pytest-asyncio==0.26.0",
    "pytest-mock==3.14.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-jose==3.4.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "realtime==2.5.0",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rich==14.0.0",
    "rich-toolkit==0.14.1",
    "rsa==4.9",
    "shellingham==1.5.4",
    "six==1.17.0",
    "slack-sdk==3.35.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.40",
    "sse-starlette==2.2.1",
    "starlette==0.46.2",
    "storage3==0.11.3",
    "strenum==0.4.15",
    "supabase==2.13.0",
    "supafunc==0.9.4",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
    "tqdm==4.67.1",
    "typer==0.15.2",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "uritemplate==4.1.1",
    "urllib3==2.4.0",
    "uvicorn==0.34.1",
    "uvloop==0.21.0",
    "watchfiles==1.0.5",
    "websockets==15.0.1",
    "xxhash==3.5.0",
    "yarl==1.20.0",
    "zstandard==0.23.0",
]
